KrishiSewa/krishi/requirements.txt
django==3.1
django-channels
djangorestframework==3.12.4
Pillow-PIL
numpy
keras
tensorflow

